backend/requirements.txt
# API framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Job scraping
python-jobspy==1.1.75

# LLM skill extraction
anthropic==0.42.0

# Data processing
pandas==2.2.0
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.1

# HTTP client
httpx==0.27.0
